refactor: created a separate fixtures for requests
